[ 2087508 ] Blurry icons in panel menu
LXpanel requests 18x18 icons, which is not usual icon size, and most
likely no theme will ship this. Instead it takes the next higher
available size it can find (which is 22x22 in most themes) and scales
it down to 18x18, which will look bad.

Now, set the menu to 16x16 now which is the closest size that ships with
every theme.

Suggested by infernux.


git-svn-id: https://lxde.svn.sourceforge.net/svnroot/lxde/trunk@867 3baf7735-d320-0410-8b56-b11ef3703fae